Moved code from core.__init__.py to version.py
